requirements.txt
scikit-learn
pandas
mlflow
dvc[ssh] # or other remote type
evidently
pyyaml